Fix map layer exceptions breaking client-side navigation
Unhandled MapLibre exceptions ("Style is not done loading",
"Cannot style non-existing layer") in the Advanced docs page
corrupted Svelte 5's reactive graph, preventing SvelteKit's
client-side router from completing page transitions.

- Extract LayerMarkersExample setup into a proper child component
  (LayerMarkersContent) with $effect lifecycle and isStyleLoaded guard
- Add isStyleLoaded check + style.load event fallback in
  CustomLayerContent to defer layer setup until style is ready
- Add layer existence checks in toggleLayer before setLayoutProperty

Closes #2
Closes #3
Closes #4
